moved SetParameter to ElementUtils
