+++ src/Libraries/RevitNodes/Elements/InternalUtilities/ElementUtils.cs
@@ -5,7 +5,8 @@ using System.Text;
 using Autodesk.DesignScript.Runtime;
 using Autodesk.Revit.DB;
 using RevitServices.Persistence;
-
+using Revit.GeometryConversion;
+using RevitServices.Transactions;
 
 namespace Revit.Elements.InternalUtilities
 {
@@ -103,7 +104,7 @@ namespace Revit.Elements.InternalUtilities
                     var paramType = param.Definition.ParameterType;
                     if (Element.IsConvertableParameterType(paramType))
                         result = param.AsDouble() * Revit.GeometryConversion.UnitConverter.HostToDynamoFactor(
-                            Element.ParameterTypeToUnitType(paramType));
+                            ParameterTypeToUnitType(paramType));
                     else
                         result = param.AsDouble();
                     break;
@@ -113,5 +114,111 @@ namespace Revit.Elements.InternalUtilities
 
             return result;
         }
+
+        /// <summary>
+        /// Convert Parameter value if necessary
+        /// </summary>
+        /// <param name="type"></param>
+        /// <param name="value"></param>
+        /// <returns></returns>
+        [SupressImportIntoVM]
+        private static double ConvertValue(ParameterType type, double value)
+        {
+            if (Element.IsConvertableParameterType(type))
+            {
+                return value * UnitConverter.DynamoToHostFactor(ParameterTypeToUnitType(type));
+            }
+
+            return value;
+        }
+
+        #region dynamic parameter setting methods
+
+        [SupressImportIntoVM]
+        public static void SetParameterValue(Autodesk.Revit.DB.Parameter param, double value)
+        {
+            if (param.StorageType != StorageType.Integer && param.StorageType != StorageType.Double)
+                throw new Exception(Properties.Resources.ParameterStorageNotNumber);
+
+            var valueToSet = GetConvertedParameterValue(param, value);
+
+            param.Set(valueToSet);
+        }
+
+        [SupressImportIntoVM]
+        public static void SetParameterValue(Autodesk.Revit.DB.Parameter param, Element value)
+        {
+            if (param.StorageType != StorageType.ElementId)
+                throw new Exception(Properties.Resources.ParameterStorageNotElement);
+
+            param.Set(value.InternalElement.Id);
+        }
+
+        [SupressImportIntoVM]
+        public static void SetParameterValue(Autodesk.Revit.DB.Parameter param, int value)
+        {
+            if (param.StorageType != StorageType.Integer && param.StorageType != StorageType.Double)
+                throw new Exception(Properties.Resources.ParameterStorageNotNumber);
+
+            var valueToSet = GetConvertedParameterValue(param, value);
+
+            param.Set(valueToSet);
+        }
+
+        [SupressImportIntoVM]
+        public static void SetParameterValue(Autodesk.Revit.DB.Parameter param, string value)
+        {
+            if (param.StorageType != StorageType.String)
+                throw new Exception(Properties.Resources.ParameterStorageNotString);
+
+            param.Set(value);
+        }
+
+        [SupressImportIntoVM]
+        public static void SetParameterValue(Autodesk.Revit.DB.Parameter param, bool value)
+        {
+            if (param.StorageType != StorageType.Integer)
+                throw new Exception(Properties.Resources.ParameterStorageNotInteger);
+
+            param.Set(value == false ? 0 : 1);
+        }
+
+        [SupressImportIntoVM]
+        public static double GetConvertedParameterValue(Autodesk.Revit.DB.Parameter param, double value)
+        {
+            var paramType = param.Definition.ParameterType;
+
+            if (Element.IsConvertableParameterType(paramType))
+            {
+                return value * UnitConverter.DynamoToHostFactor(ParameterTypeToUnitType(paramType));
+            }
+
+            return value;
+        }
+
+        [SupressImportIntoVM]
+        internal static UnitType ParameterTypeToUnitType(ParameterType parameterType)
+        {
+            switch (parameterType)
+            {
+                case ParameterType.Length:
+                    return UnitType.UT_Length;
+                case ParameterType.Area:
+                    return UnitType.UT_Area;
+                case ParameterType.Volume:
+                    return UnitType.UT_Volume;
+                case ParameterType.Angle:
+                    return UnitType.UT_Angle;
+                case ParameterType.Slope:
+                    return UnitType.UT_Slope;
+                case ParameterType.Currency:
+                    return UnitType.UT_Currency;
+                case ParameterType.MassDensity:
+                    return UnitType.UT_MassDensity;
+                default:
+                    throw new Exception(Properties.Resources.UnitTypeConversionError);
+            }
+        }
+        #endregion
     }
 }
